✨ feat: Add views for Maquina IHM and Maquina Info with data processing
- Implemented MaquinaIHMViewSet for displaying and editing IHM machine information.
- Added MaquinaInfoViewSet and MaqInfoHourProductionViewSet for machine information and hourly production data.
- Created QualidadeIHMViewSet for handling quality data from IHM machines.
- Introduced QualProdViewSet for managing qualified products.
- Developed MaquinaInfoProductionViewSet for retrieving machine information filtered by time period.
- Added StockOnCFViewSet and StockStatusViewSet for stock information from Protheus.
- Implemented CartCountViewSet for counting carts within a specified period.
- Enhanced error handling and logging across views.
- Removed obsolete maquina_info_views.py file.
- Updated SQL queries for better performance and clarity.
- Fixed minor issues in router configuration for PostgreSQL foreign tables.
