+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
   "cSpell.words": [
+    "absenteismo",
     "apscheduler",
     "asof",
     "astype",
@@ -18,6 +19,7 @@
     "downcasting",
     "dropna",
     "DTRPBG",
+    "Eficiencia",
     "Emissao",
     "ffill",
     "fillna",
@@ -27,16 +29,20 @@
     "isin",
     "iterrows",
     "jobstores",
+    "manusis",
     "MANUSIS",
     "maquinaihm",
     "maquinainfo",
     "notna",
     "NRRPET",
+    "omron",
     "Omron",
     "paes",
     "PINSP",
     "postgresql",
     "producao",
+    "protheus",
+    "Protheus",
     "pylogix",
     "PYLOGIX",
     "qualidadeihm",
