+++ backend/sfm/routers.py
@@ -20,7 +20,7 @@ class SpecificTablesRouter:
         "analysis_absent",
         "analysis_presence",
         "analysis_actionPlan",
-        "detector_metais",
+        "detector_metal",
     ]
 
     # Foreign tables do PostgreSQL para manutenção - não pode ser alterado
